fix for floating point numbers

+++ core/hosts/__init__.py
@@ -166,10 +166,10 @@ class Host:
             seconds = 0.0
         if self.vid_url:
             video = Video(self.vid_url)
-            image, seconds = await video.extract_frame(seconds=abs(int(seconds)))
+            image, seconds = await video.extract_frame(seconds=abs(float(seconds)))
         elif self.gif_url:
             gif = Gif(self.gif_url)
-            image, seconds = await gif.extract_frame(seconds=abs(int(seconds)))
+            image, seconds = await gif.extract_frame(seconds=abs(float(seconds)))
         elif self.img_url:
             image = await get_img_from_url(self.img_url)
         return image, seconds
